+++ CHANGELOG.md
@@ -5,11 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add parameter `add_union_type` to `Params.to_json`.
+
 ### Changed
 - Trainers now have parameters class set through `params_cls` instead of a generic parameter.
 - `letstune.keras.KerasTrainer` calls `model.fit` with turned off Keras logging (`verbose=0`). 
 
 ### Removed
+- Remove `get_random_params` from trainers.
 - Remove `letstune.ModelParams`. Instead, use `letstune.Params` with `model_cls` parameter.
 - Remove custom Keras and Scikit-Learn trainers. Basic trainers realize the same functionality in a more clear way.
 - Remove `letstune.Metric` class. Replace it with `str`. Now metrics are always greater-is-better.